dsrs_docker_app/requirements.txt
Django==3.1.7
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.3
pytest-cov==2.8.1
pytest-django==3.9.0
pytest==5.4.1
six==1.15.0
whitenoise==5.0.1
